# server/docs/swagger.yaml
basePath: /api/mobile
definitions:
  load.MeasuresInput:
    properties:
      additionalInfo:
        allOf:
        - $ref: '#/definitions/models.AdditionalInfo'
        description: Доп. информация к измерению
      cow:
        allOf:
        - $ref: '#/definitions/models.Cow'
        description: Загружаемая корова
      downSides:
        allOf:
        - $ref: '#/definitions/models.DownSides'
        description: Недостатки
      exterior:
        allOf:
        - $ref: '#/definitions/models.Exterior'
        description: Экстерьер коровы
      measures:
        allOf:
        - $ref: '#/definitions/models.Measures'
        description: Замеры коровы
      ratings:
        allOf:
        - $ref: '#/definitions/models.Ratings'
        description: Оценки экстерьера
      weights:
        allOf:
        - $ref: '#/definitions/models.Weights'
        description: Веса использованные в расчете
    required:
    - additionalInfo
    - cow
    - downSides
    - exterior
    - measures
    - ratings
    - weights
    type: object
  models.AdditionalInfo:
    properties:
      additionalProperty1Measure:
        description: Дополнительный параметр 1 (значение в единицах измерения)
        type: string
      additionalProperty1Name:
        description: Дополнительный параметр 1 (название)
        type: string
      additionalProperty1Value:
        description: Дополнительный параметр 1 (значение в баллах)
        type: string
      additionalProperty2Measure:
        description: Дополнительный параметр 2 (значение в единицах измерения)
        type: string
      additionalProperty2Name:
        description: Дополнительный параметр 2 (название)
        type: string
      additionalProperty2Value:
        description: Дополнительный параметр 2 (значение в баллах)
        type: string
      calvingDate:
        description: Дата отела ГГГГ-ММ-ДД
        example: "2001-03-23"
        type: string
      firstMilkingDate:
        description: Дата первого доения ГГГГ-ММ-ДД
        example: "2001-03-23"
        type: string
      lactationNumber:
        description: Номер лактации (целое, беззнаковое число)
        type: integer
    required:
    - calvingDate
    - firstMilkingDate
    type: object
  models.Cow:
    properties:
      birthDate:
        description: Дата рождения ГГГГ-ММ-ЧЧ
        example: "2001-03-23"
        type: string
      bloody:
        description: Кровность
        example: 1
        type: number
      breedName:
        description: Название породы
        example: Порода
        type: string
      holdingInn:
        description: ИНН холдинга
        type: string
      holdingName:
        description: Название холдинга
        type: string
      hozName:
        description: Название хозяйства
        type: string
      inventoryNumber:
        description: Инвентарный номер коровы
        example: "1213321"
        type: string
      name:
        description: Кличка коровы
        example: Дима
        type: string
      rshnnumber:
        description: РСХН номер коровы
        example: "1323323232"
        type: string
      selecsNumber:
        description: Селекс номер коровы
        example: 98989
        type: integer
    required:
    - birthDate
    - bloody
    - breedName
    - holdingInn
    - holdingName
    - hozName
    - inventoryNumber
    - name
    - rshnnumber
    - selecsNumber
    type: object
  models.DownSides:
    properties:
      bodyCount:
        description: Туловище
        type: number
      bodyDescription:
        description: 'Список недостатков туловища разделитель: "/"'
        type: string
      limbsCount:
        description: Конечности
        type: number
      limbsDescription:
        description: 'Список недостатков конечностей разделитель: "/"'
        type: string
      milkTypeCount:
        description: Молочный тип
        type: number
      milkTypeDescription:
        description: 'Список недостатков Молочного типа. Разделитель: "/"'
        type: string
      sacrumCount:
        description: Крестец
        type: number
      sacrumDescription:
        description: 'Список недостатков крестца разделитель: "/"'
        type: string
      summary:
        description: Сводные данные о недостатках разделитель "/"
        type: string
      udderCount:
        description: Вымя
        type: number
      udderDescription:
        description: 'Список недостатков вымени разделитель: "/"'
        type: string
    type: object
  models.Exterior:
    properties:
      assessmentDate:
        description: Дата проведения оценочных мероприятий ГГГГ-ММ-ДД
        example: "2001-03-23"
        type: string
      backBoneQuality:
        description: Качество костяка (9 баллов)
        type: number
      backNippleDiameter:
        description: Диаметр задних сосков (9 баллов)
        type: number
      backNippleLocationBackView:
        description: Расположение задних сосков вид сзади (9 баллов)
        type: number
      backUdderSegmentsLocationHeight:
        description: Высота прикрепления задних долей вымени (9 баллов)
        type: number
      backUdderSegmentsWidth:
        description: Ширина задних долей вымени (9 баллов)
        type: number
      body:
        description: Туловище (100 баллов)
        type: number
      bodyDepth:
        description: Глубина туловища (9 баллов)
        type: number
      bodyType:
        description: Тип телосложения (9 баллов)
        type: number
      category:
        description: Категория (хорошо, плохо и т.д.) ХЗ зачем это надо, но мало ли
          в разных хозяйствах категории разные
        type: string
      centralLigamentDepth:
        description: Глубина центральной связки (9 баллов)
        type: number
      chestWidth:
        description: Ширина груди (9 баллов)
        type: number
      deception:
        description: Обмускульность (9 баллов)
        type: number
      fatness:
        description: Упитанность (9 баллов)
        type: number
      forelegWalk:
        description: Поступь передних ног (9 баллов)
        type: number
      frontNippleDiameter:
        description: Диаметр передних сосков (9 балов)
        type: number
      frontNippleLength:
        description: Длинна передних сосков (9 баллов)
        type: number
      frontNippleLocationBackView:
        description: Расположение передних сосков вид сзади (9 баллов)
        type: number
      frontUdderSegmentsLocation:
        description: Прикрепление передних долей вымени (9 баллов)
        type: number
      harmonyOfMovement:
        description: Гармоничность движения (9 баллов)
        type: number
      hindLegWalkBackView:
        description: Поступь задних ног вид cзади (9 баллов)
        type: number
      hindLegWalkSideView:
        description: Поступь задних ног вид сбоку (9 баллов)
        type: number
      hoofAngle:
        description: Угол копыта (9 баллов)
        type: number
      limbs:
        description: Конечности (100 баллов)
        type: number
      milkType:
        description: Молочный тип (100 баллов)
        type: number
      rating:
        description: Суммарный рейтинг (100 баллов)
        type: number
      ribsAngle:
        description: Угол наклона ребер (9 баллов)
        type: number
      sacrum:
        description: Крестец (100 баллов)
        type: number
      sacrumAngle:
        description: Угол наклона крестца (9 баллов)
        type: number
      sacrumHeight:
        description: Высота в крестце (9 баллов)
        type: number
      sacrumLength:
        description: Длина крестца (9 баллов)
        type: number
      sacrumWidth:
        description: Ширина в крестце (9 баллов)
        type: number
      topLine:
        description: Линия верха (9 баллов)
        type: number
      udder:
        description: Вымя (100 баллов)
        type: number
      udderBalance:
        description: Баланс вымени (9 баллов)
        type: number
      udderDepth:
        description: Глубина вымени (9 баллов)
        type: number
      udderVeins:
        description: Выраженность вен вымени (9 баллов)
        type: number
    required:
    - assessmentDate
    - backBoneQuality
    - backNippleDiameter
    - backNippleLocationBackView
    - backUdderSegmentsLocationHeight
    - backUdderSegmentsWidth
    - body
    - bodyDepth
    - bodyType
    - category
    - centralLigamentDepth
    - chestWidth
    - deception
    - fatness
    - forelegWalk
    - frontNippleDiameter
    - frontNippleLength
    - frontNippleLocationBackView
    - frontUdderSegmentsLocation
    - harmonyOfMovement
    - hindLegWalkBackView
    - hindLegWalkSideView
    - hoofAngle
    - limbs
    - milkType
    - rating
    - ribsAngle
    - sacrum
    - sacrumAngle
    - sacrumHeight
    - sacrumLength
    - sacrumWidth
    - topLine
    - udder
    - udderBalance
    - udderDepth
    - udderVeins
    type: object
...
  user.ChangePasswordRequest:
    properties:
      email:
        description: Почта
        example: FedorsMail@gmail.com
        type: string
      password:
        description: Пароль
        example: FedorsPassword15
        type: string
    required:
    - email
    - password
    type: object
  user.ConfirmMailRequest:
    properties:
      code:
        description: Код с эл. почты пользователя
        example: "3295"
        type: string
      userData:
        description: Зашифрованные данные пользователя из Register
        example: 12308mjkfa01jkfa_!@#
        type: string
    required:
    - code
    - userData
    type: object
  user.ConfirmPasswordRequest:
    properties:
      code:
        description: Код с эл. почты пользователя
        example: "3295"
        type: string
      userData:
        description: Зашифрованные данные пользователя из Register
        example: 12308mjkfa01jkfa_!@#
        type: string
    required:
    - code
    - userData
    type: object
  user.LoginData:
    properties:
      email:
        description: Почта пользователя
        example: user@mail.ru
        type: string
      password:
        description: Пароль пользователя
        example: samplePassword123
        type: string
    required:
    - email
    - password
    type: object
...
  user.RegisterRequest:
    properties:
      email:
        description: Почта
        example: User321@gmail.com
        type: string
      licenseNumber:
        type: string
      nameSurnamePatronymic:
        description: Фио
        example: Иванов Федор Петрович
        type: string
      password:
        description: Пароль
        example: FedorsPassword15
        type: string
    required:
    - email
    - nameSurnamePatronymic
    - password
    type: object
...
host: cowconformationapp.ru
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Сваггер сгенерирован из кода, поэтому может содержать неточности. По мере возможности они будут описаны далее
    Все даты передаются как строки
    Большая часть рутов не возвращает вложенные объекты
    Авторизация по JWT, стандартная
    Все руты, которые требуют авторизации в теле запроса ожидают заголовок Authorization, в котором лежит строка формата "Bearer {ACCESS TOKEN}"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GenMilk mobile API
  version: "1.0"
paths:
  /load/measuredData:
    post:
      description: |-
        Требует авторизации!
        Принимает данные оценки, парсит, создает запись в БД.
        В случае успеха возвращает словарь с ключем "status" и значением "ok".
        В случае ошибки возвращает словарь с ключем "error" и строкой ошибки.
        В этом проекте предлагается хранить 4 группы 100-бальных оценок (с/без недостатками + авто/пользовательские).
        В genmilk.ru хранится только одна группа 100-бальных признаков, поля MilkType, Body, Limbs ...
        в экстерьере означают те 100-бальные оценки, которые будут в базе genmilk.ru.
        Разделитель для строк, содержащих недостатки выбран из соображения о том, что URL работает с
        разделителем "/", других причин для выбора "/" нет, можно перевыбрать, это совершенно не принципиально
      parameters:
      - description: Данные измерения коровы
        in: body
        name: measures
        required: true
        schema:
          $ref: '#/definitions/load.MeasuresInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Рут загрузки данных измерения коровы
      tags:
      - Load
  /user/changePassword:
    post:
      description: |-
        Смена пользовательского пароля
        В случае успеха возвращает словарь с ключем "userData" и значением зашифрованных данных пользователя, которые нужно будет передать в confirmPassword
        В случае ошибки возвращает словарь с ключем "error" и строкой ошибки
      parameters:
      - description: Данные смены пароля пользователя
        in: body
        name: registerData
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Change password for user
      tags:
      - ChangePassword
  /user/confirmMail:
    post:
      description: |-
        Подтверждение почты пользователя. После подтверждения пользователь становится зарегистрированным
        В случае успеха возвращает словарь с ключем "status" и значением "ok"
        В случае ошибки возвращает словарь с ключем "error" и строкой ошибки
        Код 406 - неверный код подтверждения
        Код 409 - пользователь с таким email уже существует
      parameters:
      - description: Данные для подтверждения
        in: body
        name: confirmationData
        required: true
        schema:
          $ref: '#/definitions/user.ConfirmMailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "406":
          description: Not Acceptable
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
      tags:
      - Register
  /user/confirmPassword:
    post:
      description: |-
        Подтверждение смены пароля пользователя. После подтверждения у пользователя меняется пароль
        В случае успеха возвращает словарь с ключем "status" и значением "ok"
        В случае ошибки возвращает словарь с ключем "error" и строкой ошибки
      parameters:
      - description: Данные для подтверждения
        in: body
        name: confirmationData
        required: true
        schema:
          $ref: '#/definitions/user.ConfirmPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Confirm password change
      tags:
      - ChangePassword
  /user/login:
    post:
      description: |-
        Рут логина
        В случае успеха возвращает словарь с ключем "access" и значением access токена и ключем "refresh" и значение refresh токена
        В случае ошибки возвращает словарь с ключем "error" и строкой ошибки
        В случае если пользователь не найден 404
      parameters:
      - description: Почта и пароль
        in: body
        name: loginData
        required: true
        schema:
          $ref: '#/definitions/user.LoginData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - Login
...
  /user/register:
    post:
      description: |-
        Регистрация пользователя
        В случае успеха возвращает словарь с ключем "userData" и значением зашифрованных данных пользователя, которые нужно будет передать в confirmMail
        В случае ошибки возвращает словарь с ключем "error" и строкой ошибки
      parameters:
      - description: Данные регистрации пользователя
        in: body
        name: registerData
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
      tags:
      - Register
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"